dma: fix the adma coverity issue.
typecast with uint64_t to avoid overflow of the multiplication
operation with two 32 bit value.

Coverity ID: 2567940

Reviewed-on: https://git-master.nvidia.com/r/1661093
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/dma/tegra210-adma.c
@@ -733,7 +733,7 @@ static unsigned int tegra_adma_get_residue(struct tegra_adma_chan *tdc)
 	/* get transferred data count */
 	tc_transferred = ch_regs->tc - tc_remain;
 
-	tot_xfer = (tdc->tx_buf_count * ch_regs->tc) + tc_transferred;
+	tot_xfer = (uint64_t)(tdc->tx_buf_count * ch_regs->tc) + tc_transferred;
 	tot_xfer %= desc->buf_len;
 
 	return desc->buf_len - tot_xfer;
